added DA in-memory generator in data.py
Regular data augmentation (DA) seems not so straightforward with sequence samples, as we dont want to randomly rotate each frame in a single sample sequence by a different degree. -> Custom DA is implemented

Since this DA is on the fly, the get_all_sequences_in_memory() won't cut it anymore. However, the generator is much slower.
This is partly counteracted by implementing a hybrid version: Load all raw data to memory, and apply DA on the fly from there.
Locally this gave a 1.6 - 2.0 speed up as compared to running the regular generator with DA (currently the regular generator does NOT have any DA option).

NB:
I changed the code locally quite a bit, with different names. The code definitely ran, yet before uploading I altered it to fit the online terminology again, in which I have not tested wrong naming or wrong attributes or something.